+++ app/models/meta.rb
@@ -5,7 +5,12 @@ class Meta < ApplicationRecord
 
   validates :locale, presence: true
 
-  %w(title keywords description custom slug).each do |field|
+  SEO_ATTRIBUTES = %w(title keywords description custom_head slug)
+
+  SEO_ATTRIBUTES.each do |field|
     alias_attribute "seo_#{field}".to_sym, field
   end
+
+  alias_attribute :custom, :seo_custom_head
+  alias_attribute :seo_custom, :seo_custom_head
 end